Use a registerStubCommandsPass
