+++ src/Bundle/DependencyInjection/SyliusGridExtension.php
@@ -14,11 +14,8 @@ declare(strict_types=1);
 namespace Sylius\Bundle\GridBundle\DependencyInjection;
 
 use Sylius\Bundle\CurrencyBundle\SyliusCurrencyBundle;
-use Sylius\Bundle\GridBundle\Command\StubMakeGrid;
 use Sylius\Bundle\GridBundle\Grid\GridInterface;
 use Sylius\Bundle\GridBundle\SyliusGridBundle;
-use Symfony\Bundle\MakerBundle\Maker\AbstractMaker;
-use Symfony\Bundle\MakerBundle\MakerBundle;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Loader\XmlFileLoader;
@@ -60,10 +57,6 @@ final class SyliusGridExtension extends Extension
         $container->registerForAutoconfiguration(GridInterface::class)
             ->addTag('sylius.grid')
         ;
-
-        if (!$this->isMakerEnabled($container)) {
-            $container->register(StubMakeGrid::class)->addTag('console.command');
-        }
     }
 
     public function getConfiguration(array $config, ContainerBuilder $container): Configuration
@@ -73,17 +66,5 @@ final class SyliusGridExtension extends Extension
         $container->addObjectResource($configuration);
 
         return $configuration;
-    }
-
-    private function isMakerEnabled(ContainerBuilder $container): bool
-    {
-        if (!class_exists(MakerBundle::class)) {
-            return false;
-        }
-
-        /** @var array $bundles */
-        $bundles = $container->getParameter('kernel.bundles');
-
-        return in_array(MakerBundle::class, $bundles, true);
     }
 }
